+++ seattle-soccer-tournament-backend/controllers/team.js
@@ -28,35 +28,42 @@ const show = (req, res) => {
 }
 // add new team
 const create = (req, res) => {
-    db.Team.create(req.body, (err, saveTeam) => {
-        if(err) console.log('Error in creating a team', err)
-
-        res.json({ team: saveTeam })
-    })
+    if(req.session.loggedIn) {
+        db.Team.create(req.body, (err, saveTeam) => {
+            if(err) console.log('Error in creating a team', err)
+    
+            res.json({ team: saveTeam })
+        })
+    }
 }
+
 // update team
 const update = (req, res) => {
-    db.Team.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedTeam) => {
-        if(err) console.log('Error in updating team:', err)
-
-        res.json({
-            team: updatedTeam,
-            message: "Update was successfull"
+    if(req.session.loggedIn) {
+        db.Team.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedTeam) => {
+            if(err) console.log('Error in updating team:', err)
+    
+            res.json({
+                team: updatedTeam,
+                message: "Update was successfull"
+            })
         })
-    })
+    }
 }
 
 // delete team
 const destroy = (req, res) => {
-    console.log('this is the params id', req.params.id)
-    db.Team.findByIdAndDelete(req.params.id, (err, deleteTeam) => {
-        if(err) console.log('Error in deleting team:', err)
-        console.log(deleteTeam)
-
-        res.json({
-            message: 'Team was deleted successfully!'
+    if(req.session.loggedIn){
+        console.log('this is the params id', req.params.id)
+        db.Team.findByIdAndDelete(req.params.id, (err, deleteTeam) => {
+            if(err) console.log('Error in deleting team:', err)
+            console.log(deleteTeam)
+    
+            res.json({
+                message: 'Team was deleted successfully!'
+            })
         })
-    })
+    }
 }
 
 module.exports = {
