+++ seattle-soccer-tournament-backend/controllers/tournament.js
@@ -26,26 +26,32 @@ const show = (req, res) => {
 }
 
 const create = (req, res) => {
-  db.Tournament.create(req.body, (err, saveTournament) => {
-    if(err) console.log('Error in tournament create', err)
-
-    // sending new tournament to the database
-    res.json({tournament: saveTournament})
-  })
+  if(req.session.loggedIn) {
+    db.Tournament.create(req.body, (err, saveTournament) => {
+      if(err) console.log('Error in tournament create', err)
+  
+      // sending new tournament to the database
+      res.json({tournament: saveTournament})
+    })
+  }
 }
 
 const update = (req, res) => {
-  db.Tournament.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedTournament) => {
-    if(err) console.log('Error in updating tournament: ', err)
-
-    // send updated tournament as a response
-    res.json({
-      tournament: updatedTournament,
-      message: "Update was successfull"
+  if(req.session.loggedIn) {
+    db.Tournament.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedTournament) => {
+      if(err) console.log('Error in updating tournament: ', err)
+  
+      // send updated tournament as a response
+      res.json({
+        tournament: updatedTournament,
+        message: "Update was successfull"
+      })
     })
-  })
+  }
 }
-  const destroy = (req, res) => {
+
+const destroy = (req, res) => {
+  if(req.session.loggedIn) {
     db.Tournament.findByIdAndDelete(req.params.id, (err, deleteTournament) => {
       if(err) console.log('Error in deleting tournament', err);
 
@@ -53,6 +59,7 @@ const update = (req, res) => {
         message: 'tournament was deleted successfully!'
       })
     })
+  }
 }
 
 module.exports = {
